Update word_RPG
Currently in a good place, update added was the story time for the class selection and then the ability to choose a class. I made the story and the class selection as 2 different defined var simply for convenience. I wouldn't want someone making a mistake on their class selection and then having to sit through the entire story again. 

Next thing I will need to add is a way to ONLY allow a number in the heroAge section to avoid a bug that'll crash the game. Then I will make this project into a folder and then start making other files so I can create and import things like attacks, stats, enemies, dice rolls, etc! Super excited with day 1 progress!